requirements.txt
ultralytics
opencv-python-headless
Pillow
numpy==1.23.5
opencv-python==4.9.0.80
pandas==2.2.0
PyYAML==6.0
supervision
torch
asyncio
djitellopy
flask 
flask-socketio
websocket-client